refactor: racing.js 함수 분리
