+++ src/dom.js
@@ -1,1 +1,2 @@
 export const $ = (selector) => document.querySelector(selector);
+export const $$ = (selector) => document.querySelectorAll(selector);
